+++ dev/components/DemoAlert.vue
@@ -21,11 +21,7 @@ export default {
   }),
   methods: {
     alert() {
-      // custom method name
-      const __ = this.$gettext.bind(this);
-      // use jsExtractorOpts in gettext.config.js
-      // to custom keywords
-      const msg = __("Good bye!");
+      const msg = this.__("Good bye!");// use jsExtractorOpts in gettext.config.js to extract custom keywords
       return window.alert(msg);
     },
     decrease() {
@@ -36,8 +32,7 @@ export default {
       this.n += 1;
     },
     alertPlural(n) {
-      const _n = this.$ngettext.bind(this);
-      const msg = _n("%{ n } car", "%{ n } cars", n, { n });
+      const msg = this._n("%{ n } car", "%{ n } cars", n, { n });
       return window.alert(msg);
     },
   },
