+++ dev/components/DemoDateFormat.vue
@@ -5,12 +5,12 @@
 </template>
 
 <script>
+import { useGettext } from "/@gettext/";
 import { format } from "date-fns";
 import en from "date-fns/locale/en-US";
 import fr from "date-fns/locale/fr";
 import it from "date-fns/locale/it";
 import zh from "date-fns/locale/zh-CN";
-import { useGettext } from "/@gettext/";
 
 export default {
   data: () => ({
